+++ lib/app.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:my_app/models/auth_model.dart';
-import 'package:my_app/models/editing_song_model.dart';
-import 'package:my_app/models/metronome_bpm_model.dart';
-import 'package:my_app/models/metronome_timer_model.dart';
+import 'package:my_app/models/editing_song.dart';
+import 'package:my_app/models/metronome_model.dart';
 import 'package:my_app/pages/songs_list.dart';
 import 'package:provider/provider.dart';
 
@@ -16,11 +15,8 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MultiProvider(
         providers: [
-          ChangeNotifierProvider<MetronomeBpmModel>(
-            create: (_) => MetronomeBpmModel(),
-          ),
-          ChangeNotifierProvider<MetronomeTimerModel>(
-            create: (_) => MetronomeTimerModel(),
+          ChangeNotifierProvider<MetronomeModel>(
+            create: (_) => MetronomeModel(),
           ),
           ChangeNotifierProvider<AuthModel>(
             create: (_) => AuthModel(udid),
