--- lib/models/metronome_bpm_model.dart
@@ -1,102 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:my_app/models/metronome_timer_model.dart';
-
-class MetronomeBpmModel extends ChangeNotifier {
-  int _tempoCount;
-  get tempoCount => _tempoCount;
-
-  set tempoCount(int bpm) {
-    if (bpm < 30)
-      _tempoCount = 30;
-    else if (bpm > 300)
-      _tempoCount = 300;
-    else
-      _tempoCount = bpm;
-  }
-
-  bool _isPlaying = false;
-  get isPlaying => _isPlaying;
-
-  DateTime _bpmTapStartTime;
-  var _bpmCalculateList = <int>[];
-
-  int _bpmTapCount = 0;
-  get bpmTapCount => _bpmTapCount;
-  String _bpmTapText = "TAPで計測開始";
-  get bpmTapText => _bpmTapText;
-
-  void increment() {
-    if (_tempoCount < 300) {
-      _tempoCount++;
-    }
-    notifyListeners();
-  }
-
-  void decrement() {
-    if (_tempoCount > 30) {
-      _tempoCount--;
-    }
-    notifyListeners();
-  }
-
-  void switchPlayStatus() {
-    _isPlaying = !_isPlaying;
-    notifyListeners();
-  }
-
-  void forceStop() {
-    MetronomeTimerModel().metronomeClear();
-    _isPlaying = false;
-    MetronomeTimerModel().makeMetronomeContainerStatusDefault();
-    notifyListeners();
-  }
-
-  void changeSlider(double _slideValue) {
-    _tempoCount = _slideValue.toInt();
-    notifyListeners();
-  }
-
-  void bpmTapDetector() {
-    const bpmMin = 30;
-    const bpmMax = 300;
-    const milliSeconds = 60000;
-
-    if (_bpmTapCount == 0) {
-      _bpmTapStartTime = DateTime.now();
-      _bpmTapCount++;
-      _bpmTapText = "BPM計測中...";
-      notifyListeners();
-    } else if (_bpmTapCount % 5 != 0) {
-      var _bpmDetectNow = DateTime.now();
-      var _bpmDetectDiff =
-          _bpmDetectNow.difference(_bpmTapStartTime).inMilliseconds;
-      _bpmCalculateList.add(_bpmDetectDiff);
-      _bpmTapStartTime = _bpmDetectNow;
-      _bpmTapCount++;
-      if (_bpmTapCount == 5) {
-        _bpmTapText = "計測終了";
-        notifyListeners();
-      }
-    } else {
-      int _bpmCalculateAverage = _bpmCalculateList.reduce(
-              (_bpmDiffValue, _bpmDiffElement) =>
-                  _bpmDiffValue + _bpmDiffElement) ~/
-          _bpmCalculateList.length;
-      _tempoCount = (milliSeconds / _bpmCalculateAverage).floor();
-      if (_tempoCount < bpmMin) {
-        _tempoCount = bpmMin;
-      } else if (_tempoCount > bpmMax) {
-        _tempoCount = bpmMax;
-      }
-      print("$_bpmCalculateList");
-      resetBpmTapCount();
-      notifyListeners();
-    }
-  }
-
-  void resetBpmTapCount() {
-    _bpmTapCount = 0;
-    _bpmCalculateList = <int>[];
-    _bpmTapText = "TAPで計測開始";
-  }
-}
