+++ lib/pages/custom_keyboard.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:my_app/models/editing_song_model.dart';
+import 'package:my_app/models/editing_song.dart';
 import 'package:provider/provider.dart';
 
 class CustomKeyboard extends StatelessWidget {
