+++ lib/pages/detail_page/metronome_container.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../models/metronome_bpm_model.dart';
-import '../../models/metronome_timer_model.dart';
+import '../../models/metronome_model.dart';
 
 class MetronomeContainer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Consumer<MetronomeTimerModel>(builder: (_, model, __) {
+    return Consumer<MetronomeModel>(builder: (_, model, __) {
       return Row(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
@@ -31,23 +30,22 @@ class MetronomeContainerWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      width: 20,
-      height: 20,
-      decoration: BoxDecoration(
-          border: Border.all(
-            color: Colors.black,
-            width: 0.1,
-          ),
-          shape: BoxShape.circle,
-          color: !Provider.of<MetronomeBpmModel>(context).isPlaying
-              ? Colors.white
-              : (Provider.of<MetronomeTimerModel>(context)
-                              .metronomeContainerStatus %
-                          contentNum ==
-                      contentState
-                  ? Colors.orange
-                  : Colors.white)),
-    );
+    return Consumer<MetronomeModel>(builder: (_, model, __) {
+      return Container(
+        width: 20,
+        height: 20,
+        decoration: BoxDecoration(
+            border: Border.all(
+              color: Colors.black,
+              width: 0.1,
+            ),
+            shape: BoxShape.circle,
+            color: !model.isPlaying
+                ? Colors.white
+                : (model.metronomeContainerStatus % contentNum == contentState
+                    ? Colors.orange
+                    : Colors.white)),
+      );
+    });
   }
 }
