+++ lib/pages/detail_page/settings_drawer.dart
@@ -1,12 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
-import '../../models/metronome_timer_model.dart';
+import '../../models/metronome_model.dart';
 
-Drawer settingsDrawer(
-    BuildContext context, int bpm, String title, String docId) {
+Drawer settingsDrawer(BuildContext context, MetronomeModel model, int bpm,
+    String title, String docId) {
   final double titleTextFont = 16;
   final insertPadding = Padding(padding: EdgeInsets.all(10));
 
@@ -48,8 +47,7 @@ Drawer settingsDrawer(
           ),
           ListTile(
               tileColor: Theme.of(context).primaryColorDark,
-              title: Text(
-                  Provider.of<MetronomeTimerModel>(context).metronomeSound,
+              title: Text(model.metronomeSound,
                   style: TextStyle(color: Colors.white)),
               onTap: () {
                 print("Metronome Sound");
@@ -65,11 +63,7 @@ Drawer settingsDrawer(
           ),
           ListTile(
               tileColor: Theme.of(context).primaryColorDark,
-              title: Text(
-                  "回数：" +
-                      Provider.of<MetronomeTimerModel>(context)
-                          .countInTimes
-                          .toString(),
+              title: Text("回数：" + model.countInTimes.toString(),
                   style: TextStyle(color: Colors.white)),
               onTap: () {
                 print("Count-in Times");
